requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
markitdown[docx,pptx,xlsx]
pypdf
python-docx
openai
python-dotenv
tiktoken